+++ cyclenet/src/main/java/com/ingesoft/cyclenet/logic/CasosDeUsoUsuarios.java
@@ -1,18 +1,12 @@
 package com.ingesoft.cyclenet.logic;
 
-import java.util.ArrayList;
 import java.util.Optional;
-import java.time.LocalDate;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.sql.Date;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ingesoft.cyclenet.dataAccess.RepositorioPublicacion;
 import com.ingesoft.cyclenet.dataAccess.RepositorioUsuario;
-import com.ingesoft.cyclenet.domain.Publicacion;
 import com.ingesoft.cyclenet.domain.Usuario;
 
 @Service
@@ -21,8 +15,7 @@ public class CasosDeUsoUsuarios {
     @Autowired
     protected RepositorioUsuario repositorioUsuario;
 
-    @Autowired
-    protected RepositorioPublicacion repositorioPublicacion;
+    
     
     public void iniciarSesion(String login, String password) throws Exception{
 
@@ -52,30 +45,6 @@ public class CasosDeUsoUsuarios {
         return;
     }
 
-    
-
-    public void subirPublicacion(Usuario usuario, String mensaje, Boolean foto, Boolean lugar){
-        //Guardar publicación
-        Date fecha = Date.valueOf(LocalDate.now());
-        Publicacion publicacion = new Publicacion(mensaje, foto, lugar, fecha, usuario);
-        usuario.getPublicaciones().add(publicacion);
-        repositorioPublicacion.save(publicacion);
-
-        if(lugar == true){
-            mostrarPublicacionLugar(publicacion);
-            return;
-        }
-
-        //Mostrar publicacion
-         //Mostrar texto, e imagen
-         return;
-    }
-
-    public void mostrarPublicacionLugar(Publicacion publicacion){
-        //Procedimiento para obtener y mostrar el lugar
-        return;
-    }
-
     public void seguirUsuario(){
 
     }
@@ -89,7 +58,7 @@ public class CasosDeUsoUsuarios {
     }
 
     public void realizarCalificacion(){
-        
+
     }
